add scala3 support for s3 (#167)
* add futiles license

* some more compile issues

* more compile issues

* Update S3.scala

* try .toInt

* remove ASF header from futiles code

* Final touches to get S3 working for Scala 3

---------

+++ s3/src/test/scala/org/apache/pekko/stream/connectors/s3/impl/auth/SplitAfterSizeWithContextSpec.scala
@@ -79,6 +79,7 @@ class SplitAfterSizeWithContextSpec(_system: ActorSystem)
         Seq((ByteString(17, 18), 2))))
   }
 
-  def bytes(start: Byte, end: Byte): Array[Byte] = (start to end).map(_.toByte).toArray[Byte]
+  // https://github.com/lampepfl/dotty/issues/18068
+  def bytes(start: Byte, end: Byte): Array[Byte] = (start.toInt to end).map(_.toByte).toArray[Byte]
 
 }
